Move blog tags to blog app
