+++ app/templates/base.html
@@ -1,4 +1,4 @@
-{% load static wagtailcore_tags wagtailuserbar navigation_tags %}
+{% load static wagtailcore_tags wagtailuserbar blog_tags %}
 
 <!DOCTYPE html>
 <html lang="en">
